+++ docs/conf.py
@@ -27,7 +27,7 @@ author = u'Chris Müller'
 # The short X.Y version
 version = u'2.0'
 # The full version, including alpha/beta/rc tags
-release = u'2.0.0-dev'
+release = u'2.0.0'
 
 
 # -- General configuration ---------------------------------------------------