+++ docs/changelog.rst
@@ -8,8 +8,17 @@ All notable changes to this project will be documented in this file.
 The format is based on `Keep a Changelog <https://keepachangelog.com/en/1.0.0/>`_\ , and this project adheres
 to `Semantic Versioning <https://semver.org/spec/v2.0.0.html>`_.
 
-`Unreleased <https://github.com/brotkrueml/jobrouter-client/compare/v1.4.0...HEAD>`_
+`Unreleased <https://github.com/brotkrueml/jobrouter-client/compare/v2.0.0...HEAD>`_
 ----------------------------------------------------------------------------------------
+
+`2.0.0 <https://github.com/brotkrueml/jobrouter-client/compare/v1.4.0...v2.0.0>`_ - 2023-02-28
+--------------------------------------------------------------------------------------------------
+
+Added
+^^^^^
+
+
+* Compatibility with JobRouter® 2023.1
 
 Changed
 ^^^^^^^